requirements.txt
asgiref==3.4.1
bson==0.5.8
dataclasses==0.6
dj-database-url==0.5.0
Django==3.2.9
django-admin==2.0.1
django-base-url==2020.12.3
django-cors-headers==3.10.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-filter==21.1
django-six==1.0.4
djangorestframework==3.12.4
djongo==1.3.6
gunicorn==20.1.0
importlib-metadata==4.8.2
Markdown==3.3.5
mongoengine==0.23.1
Pillow==8.4.0
pymongo==3.12.1
python-dateutil==2.8.2
pytz==2021.3
screen==1.0.1
six==1.16.0
sqlparse==0.2.4
xlwt==1.3.0
zipp==3.6.0
